feat(global): preferences settings

+++ packages/web-extension/src/background/index.ts
@@ -19,7 +19,7 @@ chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
 })
 
 chrome.runtime.onInstalled.addListener(function(details) {
-    if (details.reason == "update") {
+    if (details.reason == "update" && (details.previousVersion && details.previousVersion < "0.4.1")) {
       chrome.storage.local.clear();
     }
   });
